HW3_node/Submit/requirement.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.6.0=py310hca03da5_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
blas=1.0=openblas
bleach=6.1.0=pypi_0
bzip2=1.0.8=h80987f9_6
ca-certificates=2024.3.11=hca03da5_0
certifi=2024.2.2=py310hca03da5_0
cffi=1.16.0=py310h80987f9_1
charset-normalizer=3.3.2=pypi_0
dgl=2.2.1=pypi_0
filelock=3.13.1=py310hca03da5_0
frozenlist=1.4.1=pypi_0
fsspec=2024.3.1=py310hca03da5_0
gmp=6.2.1=hc377ac9_3
gmpy2=2.1.2=py310h8c48613_0
idna=3.7=pypi_0
jinja2=3.1.3=py310hca03da5_0
joblib=1.4.2=pypi_0
kaggle=1.6.12=pypi_0
libcxx=14.0.6=h848a8c0_0
libffi=3.4.4=hca03da5_1
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libopenblas=0.3.21=h269037a_0
libprotobuf=3.20.3=h514c7bf_0
libuv=1.44.2=h80987f9_0
llvm-openmp=14.0.6=hc6e5704_0
markupsafe=2.1.3=py310h80987f9_0
mpc=1.1.0=h8c48613_1
mpfr=4.0.2=h695f6f0_1
mpmath=1.3.0=py310hca03da5_0
multidict=6.0.5=pypi_0
ncurses=6.4=h313beb8_0
networkx=3.1=py310hca03da5_0
ninja=1.10.2=hca03da5_5
ninja-base=1.10.2=h525c30c_5
numpy=1.26.4=py310h3b2db8e_0
numpy-base=1.26.4=py310ha9811e2_0
openssl=1.1.1w=h1a28f6b_0
pandas=2.2.2=pypi_0
pip=24.0=py310hca03da5_0
psutil=5.9.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=2.5.3=py310hca03da5_0
pydantic-core=2.14.6=py310hf0e4da2_0
pyparsing=3.1.2=pypi_0
python=3.10.0=hbdb9e5c_5
python-dateutil=2.9.0.post0=pypi_0
python-slugify=8.0.4=pypi_0
pytorch=2.2.0=gpu_mps_py310h87e4ab7_100
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h1a28f6b_0
requests=2.31.0=pypi_0
scikit-learn=1.4.2=pypi_0
scipy=1.13.0=pypi_0
setuptools=69.5.1=py310hca03da5_0
six=1.16.0=pypi_0
sleef=3.5.1=h80987f9_2
sqlite=3.45.3=h80987f9_0
sympy=1.12=py310hca03da5_0
text-unidecode=1.3=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.14=h6ba3021_0
torch-geometric=2.5.3=pypi_0
torchdata=0.7.1=pypi_0
tqdm=4.66.4=pypi_0
typing-extensions=4.11.0=py310hca03da5_0
typing_extensions=4.11.0=py310hca03da5_0
tzdata=2024.1=pypi_0
urllib3=2.2.1=pypi_0
webencodings=0.5.1=pypi_0
wheel=0.43.0=py310hca03da5_0
xz=5.4.6=h80987f9_1
yaml=0.2.5=h1a28f6b_0
yarl=1.9.4=pypi_0
zlib=1.2.13=h18a0788_1
